big fixes
fixed the way the dlmread file reads.. seems to read the first column as zeros (must be how asc is made in qgis) so i changed ti to read from 1. and i chaged the rows to start at 6. idk it must be a new matlab version thing.

then i changed the easting and northing dficion because of similar reasons.

then i added a cleaning to the exposrted data to delete values where wavelength was nan. its not worth dealing with the nans for now so it lessens the results of height and leesdie angle but alteast you get the full measurements.

for threshold, i changes it to be less than 0.1.. htis always changes so this is not a big deal here.

the major thing i did in findWL is i had commented the trimming (because i was working with really nice rectangular and full grids) and i forgot that with REAL data it really changes the size of hte grids and that is why the trimming is needed. so i figured that out by looking at formativeWL size and it was almost double the size of the meanleethresh. so i jus tuncommented out the trimming section (lines 46-63.) and it worked perfectly.

hopefully no more indexing problems!!

if tehre are some start with how ascii is being read in dlmread.. i sense global mapper and  qgis save these formats differently.